# openapi.yml
openapi: 3.0.0
components:
  schemas:
    PutDocumentRequest:
      properties:
        SomeObject:
          type: object
          properties:
            SomeAttribute:
              type: string
  securitySchemes: {}
info:
  title: ''
  description: ''
  version: 2328e44b-673b-4826-b7b4-d3540a57086e
paths:
  //:
    get:
      operationId: bienvenido
      summary: Create User
      description: Creates a user and then sends a generated password email
      responses:
        '201':
          description: A user object along with generated API Keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PutDocumentResponse'
        '500':
          description: An error message when creating a new user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  //buscar/planetas:
    post:
      operationId: buscarPlanetas
      summary: Create User
      description: Creates a user and then sends a generated password email
      responses:
        '201':
          description: A user object along with generated API Keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PutDocumentResponse'
        '500':
          description: An error message when creating a new user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  //registrar/personaje:
    post:
      operationId: registrarPersonaje
      summary: Create User
      description: Creates a user and then sends a generated password email
      responses:
        '201':
          description: A user object along with generated API Keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PutDocumentResponse'
        '500':
          description: An error message when creating a new user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  //mostrar/personajes:
    get:
      operationId: mostrarPersonaje
      summary: Create User
      description: Creates a user and then sends a generated password email
      responses:
        '201':
          description: A user object along with generated API Keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PutDocumentResponse'
        '500':
          description: An error message when creating a new user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'